LayoutTests/media/modern-media-controls/airplay-support/airplay-support-expected.txt

Testing the AirPlaySupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Making AirPlay routes available.
PASS mediaController.controls.airplayButton.enabled is true
PASS mediaController.controls.airplayButton.on is false
PASS mediaController.controls.muteButton.enabled is true
PASS mediaController.controls.autoHideController.fadesWhileIdle is true

Entering AirPlay playback mode.
PASS mediaController.controls.airplayButton.enabled is true
PASS mediaController.controls.airplayButton.on is true
PASS mediaController.controls.muteButton.enabled is false
PASS mediaController.controls.autoHideController.fadesWhileIdle is false

PASS successfullyParsed is true

TEST COMPLETE
